eahw2697 moved setting of finish time text to editShift
